api/requirements.txt
Django>=2.0.3
djangorestframework>=3.7.7
